Also accept strings with file URIs in read(), write() and append()

+++ src/main/php/io/Files.class.php
@@ -16,31 +16,34 @@ class Files {
    * $str= Files::read(new File('/etc/passwd'));
    * ```
    *
-   * @param  io.File $file
+   * @param  string|io.File $file
    * @return string file contents
    * @throws io.IOException
    * @throws io.FileNotFoundException
    */
   public static function read($file) {
-    if ($file->isOpen()) {
-      $bytes= '';
-      do {
-        $bytes.= $file->read();
-      } while (!$file->eof());
-    } else {
-      clearstatcache();
-      $file->open(File::READ);
-      $size= $file->size();
+    if ($file instanceof File) {
+      if ($file->isOpen()) {
+        $bytes= '';
+        do {
+          $bytes.= $file->read();
+        } while (!$file->eof());
+      } else {
+        clearstatcache();
+        $file->open(File::READ);
+        $size= $file->size();
 
-      // Read until EOF. Best case scenario is that this will run exactly once.
-      $bytes= '';
-      do {
-        $l= $size - strlen($bytes);
-        $bytes.= $file->read($l);
-      } while ($l > 0 && !$file->eof());
-      $file->close();
+        // Read until EOF. Best case scenario is that this will run exactly once.
+        $bytes= '';
+        do {
+          $l= $size - strlen($bytes);
+          $bytes.= $file->read($l);
+        } while ($l > 0 && !$file->eof());
+        $file->close();
+      }
+      return $bytes;
     }
-    return $bytes;
+    return self::read(new File($file));
   }
 
   /**
@@ -51,41 +54,47 @@ class Files {
    * $written= Files::write(new File('myfile'), 'Hello world');
    * ```
    *
-   * @param  io.File $file
+   * @param  string|io.File $file
    * @param  string $bytes
    * @return int
    * @throws io.IOException
    */
   public static function write($file, $bytes) {
-    if ($file->isOpen()) {
-      $file->seek(0, SEEK_SET);
-      $written= $file->write($bytes);
-      $file->truncate($written);
-    } else {
-      $file->open(File::WRITE);
-      $written= $file->write($bytes);
-      $file->close();
+    if ($file instanceof File) {
+      if ($file->isOpen()) {
+        $file->seek(0, SEEK_SET);
+        $written= $file->write($bytes);
+        $file->truncate($written);
+      } else {
+        $file->open(File::WRITE);
+        $written= $file->write($bytes);
+        $file->close();
+      }
+      return $written;
     }
-    return $written;
+    return self::write(new File($file), $bytes);
   }
 
   /**
    * Append file contents. If the file was previously open, it is not closed
    * after the bytes have been written.
    *
-   * @param  io.File $file
+   * @param  string|io.File $file
    * @param  string $bytes
    * @return int
    * @throws io.IOException
    */
   public static function append($file, $bytes) {
-    if ($file->isOpen()) {
-      $written= $file->write($bytes);
-    } else {
-      $file->open(File::APPEND);
-      $written= $file->write($bytes);
-      $file->close();
+    if ($file instanceof File) {
+      if ($file->isOpen()) {
+        $written= $file->write($bytes);
+      } else {
+        $file->open(File::APPEND);
+        $written= $file->write($bytes);
+        $file->close();
+      }
+      return $written;
     }
-    return $written;
+    return self::append(new File($file), $bytes);
   }
 }
